cleanup and example
