+++ src/evidently/metrics/regression_performance/predicted_vs_actual.py
@@ -1,24 +1,20 @@
 import json
-from typing import Dict
 from typing import List
 from typing import Optional
-from typing import Tuple
-from typing import Type
-from typing import TypeVar
 from typing import Union
-from typing import overload
 
 import numpy as np
 from plotly import graph_objs as go
-from typing_extensions import Literal
 
 from evidently.base_metric import InputData
 from evidently.base_metric import Metric
 from evidently.base_metric import MetricResult
 from evidently.core import IncludeTags
+from evidently.metric_results import raw_agg_properties
 from evidently.metrics.regression_performance.objects import PredActualScatter
 from evidently.metrics.regression_performance.objects import scatter_as_dict
 from evidently.model.widget import BaseWidgetInfo
+from evidently.options.base import AnyOptions
 from evidently.renderers.base_renderer import MetricRenderer
 from evidently.renderers.base_renderer import default_renderer
 from evidently.renderers.html_widgets import header_text
@@ -26,54 +22,16 @@ from evidently.utils.data_operations import process_columns
 from evidently.utils.visualizations import plot_scatter
 
 
-class RegressionPredictedVsActualScatterResults(MetricResult):
+class AggPredActualScatter(MetricResult):
     class Config:
         dict_include = False
         tags = {IncludeTags.Render}
 
-    current: PredActualScatter
-    reference: Optional[PredActualScatter]
+    # todo
+    data: dict
 
 
-AggPredActualScatter = dict
-
-TR = TypeVar("TR")
-TA = TypeVar("TA")
-
-
-@overload
-def raw_agg_properties(
-    field_name, raw_type: Type[TR], agg_type: Type[TA], optional: Literal[True]
-) -> Tuple[Optional[TR], Optional[TA]]:
-    ...
-
-
-@overload
-def raw_agg_properties(field_name, raw_type: Type[TR], agg_type: Type[TA], optional: Literal[False]) -> Tuple[TR, TA]:
-    ...
-
-
-def raw_agg_properties(field_name, raw_type: Type[TR], agg_type: Type[TA], optional: bool) -> Tuple[TR, TA]:
-    def property_raw(self):
-        val = getattr(self, field_name)
-        if optional and val is None:
-            return None
-        if not isinstance(val, raw_type):
-            raise ValueError("Raw data not available")
-        return val
-
-    def property_agg(self):
-        val = getattr(self, field_name)
-        if optional and val is None:
-            return None
-        if not isinstance(val, agg_type):
-            raise ValueError("Agg data not available")
-        return val
-
-    return property(property_raw), property(property_agg)  # type: ignore[return-value]
-
-
-class RegressionPredictedVsActualScatterResults2(MetricResult):
+class RegressionPredictedVsActualScatterResults(MetricResult):
     class Config:
         dict_include = False
         tags = {IncludeTags.Render}
@@ -85,22 +43,11 @@ class RegressionPredictedVsActualScatterResults2(MetricResult):
     reference_raw, reference_agg = raw_agg_properties("reference", PredActualScatter, AggPredActualScatter, True)
 
 
-class RegressionPredictedVsActualScatterResults3(MetricResult):
-    class Config:
-        dict_include = False
-        tags = {IncludeTags.Render}
+class RegressionPredictedVsActualScatter(Metric[RegressionPredictedVsActualScatterResults]):
+    def __init__(self, options: AnyOptions = None):
+        super().__init__(options=options)
 
-    current: Optional[PredActualScatter]
-    current_agg: Optional[AggPredActualScatter]
-    reference: Optional[PredActualScatter]
-    reference_agg: Optional[AggPredActualScatter]
-
-
-class RegressionPredictedVsActualScatter(Metric[RegressionPredictedVsActualScatterResults2]):
-    def __init__(self, agg_data: bool = False):
-        self.agg_data = agg_data
-
-    def calculate(self, data: InputData) -> RegressionPredictedVsActualScatterResults2:
+    def calculate(self, data: InputData) -> RegressionPredictedVsActualScatterResults:
         dataset_columns = process_columns(data.current_data, data.column_mapping)
         target_name = dataset_columns.utility_columns.target
         prediction_name = dataset_columns.utility_columns.prediction
@@ -110,20 +57,20 @@ class RegressionPredictedVsActualScatter(Metric[RegressionPredictedVsActualScatt
             raise ValueError("The columns 'target' and 'prediction' columns should be present")
         if not isinstance(prediction_name, str):
             raise ValueError("Expect one column for prediction. List of columns was provided.")
-        if not self.agg_data or self.context.agg_data:
+        if not self.options.agg_data:
             curr_df = self._make_df_for_plot(curr_df, target_name, prediction_name, None)
             current_scatter = PredActualScatter(predicted=curr_df[prediction_name], actual=curr_df[target_name])
             reference_scatter: Optional[PredActualScatter] = None
             if data.reference_data is not None:
                 ref_df = self._make_df_for_plot(ref_df, target_name, prediction_name, None)
                 reference_scatter = PredActualScatter(predicted=ref_df[prediction_name], actual=ref_df[target_name])
-            return RegressionPredictedVsActualScatterResults2(current=current_scatter, reference=reference_scatter)
+            return RegressionPredictedVsActualScatterResults(current=current_scatter, reference=reference_scatter)
 
-        current_agg = {"a": 1, "b": 2}
+        current_agg = AggPredActualScatter(data={"a": 1, "b": 2})
         reference_agg = None
         if data.reference_data is not None:
-            reference_agg = {"a": 5, "b": 2}
-        return RegressionPredictedVsActualScatterResults2(current=current_agg, reference_agg=reference_agg)
+            reference_agg = AggPredActualScatter(data={"a": 5, "b": 2})
+        return RegressionPredictedVsActualScatterResults(current=current_agg, reference_agg=reference_agg)
 
     def _make_df_for_plot(self, df, target_name: str, prediction_name: str, datetime_column_name: Optional[str]):
         result = df.replace([np.inf, -np.inf], np.nan)
@@ -162,9 +109,10 @@ class RegressionPredictedVsActualScatterRenderer(MetricRenderer):
         ]
 
     def render_agg(self, current: AggPredActualScatter, reference: Optional[AggPredActualScatter]):
-        data = [go.Bar(x=list(current.keys()), y=list(current.values()))]
+        # todo
+        data = [go.Bar(x=list(current.data.keys()), y=list(current.data.values()))]
         if reference is not None:
-            data.append(go.Bar(x=list(reference.keys()), y=list(reference.values())))
+            data.append(go.Bar(x=list(reference.data.keys()), y=list(reference.data.values())))
         fig_agg = go.Figure(data=data)
         fig_agg = json.loads(fig_agg.to_json())
 
@@ -180,6 +128,6 @@ class RegressionPredictedVsActualScatterRenderer(MetricRenderer):
 
     def render_html(self, obj: RegressionPredictedVsActualScatter) -> List[BaseWidgetInfo]:
         result = obj.get_result()
-        if obj.agg_data or obj.context.agg_data:
+        if obj.options.agg_data:
             return self.render_agg(result.current_agg, result.reference_agg)
         return self.render_raw(result.current_raw, result.reference_raw)
